+++ CHANGELOG.md
@@ -13,11 +13,17 @@ Types of changes:
 - *Security* in case of vulnerabilities. 
 
 <!-- changelog-linker -->
-## Draft - 2021-07-16
+## Draft - 2021-07-31
+
+# Added
+- Added clickable link in the PDF footer
 
 # Changed
 - PDF preview on mobile when signing
 - Only show request signing button if file not signed
+- Associate signed file to LibreSign file
+- API: return LibreSign UUID on sign methods
+- Endpoint: /account/create/{uuid}, remove required of field signPassword
 
 ## 2.4.3 - 2021-07-14
 
@@ -46,6 +52,10 @@ Types of changes:
 - View document on mobile before sign
 - Markdown formatting for description
 - Libresign button in file options in theh Files app
+- Button to redirect to files to view the document
+- Legal information on Validation screen, configure in Admin settings
+- Validation page, validating by UUID and ID
+- Button that takes you to the validation page on all `.signed` and `.signed` files
 
 # Changed
 - Bump max Nextcloud version to 23
@@ -55,6 +65,8 @@ Types of changes:
 - Logo replaced by new logo
 - It will only verify the password if nextcloud requests confirmation of the password by the OC.
 - Check if has pfx
+- After signing the document, it will update the app files
+- Changed wizard to split user creation and pfx creation
 
 # Removed
 - Removed dsv folder
@@ -68,6 +80,7 @@ Types of changes:
 - Add ellipsis to pdf file title
 - Now it is possible to choose a file even if it is inside x folders
 - Clear uuid field before returns
+- Button to redirect to document validation page
 
 ## 2.3.0 - 2021-05-22